requirements.txt
onnxruntime
pillow
gradio
numpy
